MNT: Merge everything back in to skxray.core.correlation.py
